+++ src/fields/CountryMultiSelectField.php
@@ -47,22 +47,25 @@ class CountryMultiSelectField extends CountrySelectBaseOptionsField
     /**
      * @inheritdoc
      */
-    public function getInputHtml($value, ElementInterface $element = null): string
+    public function getInputHtml($value, ?ElementInterface $element = null): string
     {
+        $name = $this->handle;
+        $options = $this->translatedOptions();
+
         // Get our id and namespace
-        $id = Craft::$app->getView()->formatInputId($this->handle);
+        $id = Craft::$app->getView()->formatInputId($name);
         $namespacedId = Craft::$app->getView()->namespaceInputId($id);
 
         // Render the input template
         return Craft::$app->getView()->renderTemplate(
             'country-select-field/_multiselect',
             [
-                'name' => $this->handle,
+                'name' => $name,
                 'values' => $value,
                 'field' => $this,
                 'id' => $id,
                 'namespacedId' => $namespacedId,
-                'options' => $this->translatedOptions(),
+                'options' => $options,
             ]
         );
     }